Implement ReadyQueue
